Initial setup of the new go SDK
The new SDK provides support for PipelineRun and TaskRun events,
cloudevents binding and JSON binding. It include tests to verify
that the JSON produced by the SDK matches what's defined in the
spec.

There is no CLI functionality included, this can be consumed as
a library as described in the README.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>