Add jsonschema struct tags
The omitempty struct tag is interpreted by the schema generation
library like a non-required field, so we cannot set "omitempty"
on required fields, else the schema will be invalid.

To solve the issue, use instead minLenght tags for all string
required fields, to indicate that a non-empty value must be
provided for validation to pass.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>