Validate from schema as strings
Schema files are problematic in a go module as non-go files are
not included in the module so they would need to be embedded in
a go file.

Instead, use the init() of the binding module to generate all schemas
on the fly and store them as string in a map, which we use then
for validation.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>