Handle event versions in the SDK
Change CDEventType to be a struct to allow for one-time parsing
of the type string. The event type is still stored as string in
the context so that it may unmarshal without the need for a
custom unmarshaller.

When producing events, the SDK produces the latest version of an
event, the only one know to the SDK.
When consuming events, the SDK allows for versions different from
the SDK ones, as long as they are compatible with the SDK one,
according to the versioning rules from the spec.

The spec submodule has been moved into pkg/api so that we may use
go embed to embed the schemas from the spec into the event modules
which allows to embed the schema definitions in the SDK module.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>