Generate the SDK, implement v0.2.0
Add the generator tool which generates a large part of the SDK
surface from the JSON schemas defined in the spec repo.

Define a new test-only event which is used to run SDK unit tests
independent from the actual events defined in the spec.

Add a new, mostly generated, examples test. This test verifies
that the SDK can be used to generate the examples defined in the
spec repository, and excercises the code automatically generated
from the schemas.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>