Embed the schemas in generated files
go:embed includes files at build time into a binary.
It doesn't work when consuming the SDK as a go module, since
the json files are not included in the module itself.

Instead of using embed, load the json file into the template
file and embed it directly (at generate time) into the
generated go files.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>