Add support for complex objects in content
The generator script allows for content field to be either
string or "References". This extends the generator to allow
for content fields being objects, with the limitations that
object must themselves be made of string fields alone.

Templates are extended to define new structs that hold the
object types and to use those types where needed.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>