Set the CloudEvent ID to match CDEvents Id
When using the CloudEvent binding, the CloudEvent ID must match,
according to the specification, with the CDEvents Id.

This can be done today by the event producer via the CloudEvents
and CDEvents SDKs, however, since it's required by the spec,
it's better implemented as default by the CDEvents SDK directly.

Fixes: #57

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>