Add specification packages
Add a go package that holds an unversioned alias of the API for
creating events. This let user import the versioned package and
have access to a simpler API which picks the right version of
each event based on the spec version.

There is still work to be done to support multiple versions, as
the context may vary across versions, which is not yet handled.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>