Add specversion in several places
Add the spec version to the event contructors.
Use the constructors aliases in docs.go to pass the correct version of
the spec directly.

Extend the parsing API with generics, so that it may take the type of
context as input. For now, the only type available is Context, but
once we do breaking changes to the spec, there will be more than one
version of Context, and the parsing method will need to take a version
specific context type.

Change the context.type from String to CDEventsType and define custom
JSON marshaller and unmarshaller to keep the error handling logic in
the correct place in the code and simplify the code in general.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>