Add specification packages (#78)
* Add specification packages

Add a go package that holds an unversioned alias of the API for
creating events. This let user import the versioned package and
have access to a simpler API which picks the right version of
each event based on the spec version.

There is still work to be done to support multiple versions, as
the context may vary across versions, which is not yet handled.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Add specversion in several places

Add the spec version to the event contructors.
Use the constructors aliases in docs.go to pass the correct version of
the spec directly.

Extend the parsing API with generics, so that it may take the type of
context as input. For now, the only type available is Context, but
once we do breaking changes to the spec, there will be more than one
version of Context, and the parsing method will need to take a version
specific context type.

Change the context.type from String to CDEventsType and define custom
JSON marshaller and unmarshaller to keep the error handling logic in
the correct place in the code and simplify the code in general.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Adapt binding functions to multi-spec

Make the test events part of a test spec version.
Move parsing functions to spec specific packages where needed.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Fix the spec versions

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Reduce number of templates

Use a bit of template logic to reduce the number of templates.
Add no-build tags to the test SDK files.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Improve the generator file handling

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Make the embedded schema work with multiple specs

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Add some more unit tests for the generator

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

---------

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>