Add the spec-v0.4 submodule
Add the new version of the spec and begin to address the various
issues encountered with mulitple versions.

In this commit, move all the schema definitions across specs
into a single schemas.go in the api package, and adapt the existing
events template to benefit from that.

The same DB of schema files is used by the generator to load all
schemas in the beginning and resolve references through a custom
resolver.

Move the example and factory tests into the spec specific
packages. The conformance tests do not pass yet as a few extra
fields need to be set which is not possible yet because changes to
the context (including links) have not been processed yet.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>