Support for v04 Context
Add a new context type for new fields in the context introduced
in v0.4. Create new interfaces to match the new fields as well
and update the templates to use the new context and implement
the new interfaces when the spec version is not 0.3.x.

Add a new version of the test type that is used to test the
generator code so that we can test generating for the new version.

Updated the logic to generate the schema in a golang file to
use it as a local schema database for the jsonschema library
to validate documents. The schemas are precompiled at init
time now, which simplifies the logic downstream in the code.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>