Import spec v0.4 and add support for v0.4 Context (#83)
* Add the spec-v0.4 submodule

Add the new version of the spec and begin to address the various
issues encountered with mulitple versions.

In this commit, move all the schema definitions across specs
into a single schemas.go in the api package, and adapt the existing
events template to benefit from that.

The same DB of schema files is used by the generator to load all
schemas in the beginning and resolve references through a custom
resolver.

Move the example and factory tests into the spec specific
packages. The conformance tests do not pass yet as a few extra
fields need to be set which is not possible yet because changes to
the context (including links) have not been processed yet.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

* Support for v04 Context

Add a new context type for new fields in the context introduced
in v0.4. Create new interfaces to match the new fields as well
and update the templates to use the new context and implement
the new interfaces when the spec version is not 0.3.x.

Add a new version of the test type that is used to test the
generator code so that we can test generating for the new version.

Updated the logic to generate the schema in a golang file to
use it as a local schema database for the jsonschema library
to validate documents. The schemas are precompiled at init
time now, which simplifies the logic downstream in the code.

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>

---------

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>